COST mapping for synthea v2.7, v3.0 into cdm v5.3, v5.4

+++ R/DropSyntheaTables.r
@@ -19,6 +19,8 @@ DropSyntheaTables <- function(connectionDetails, syntheaSchema)
   syntheaTables <- c(
     "ALLERGIES",
     "CAREPLANS",
+	"CLAIMS",
+	"CLAIMS_TRANSACTIONS",
     "CONDITIONS",
     "DEVICES",
     "ENCOUNTERS",
@@ -28,25 +30,33 @@ DropSyntheaTables <- function(connectionDetails, syntheaSchema)
     "OBSERVATIONS",
     "ORGANIZATIONS",
     "PATIENTS",
+	"PAYERS",
+	"PAYER_TRANSITIONS",
     "PROCEDURES",
-    "PROVIDERS"
+    "PROVIDERS",
+	"SUPPLIES"
   )
 
   conn <- DatabaseConnector::connect(connectionDetails)
   allTables <- DatabaseConnector::getTableNames(conn, syntheaSchema)
-  writeLines("Dropping Synthea tables...")
   tablesToDrop <- allTables[which(allTables %in% syntheaTables)]
-  sql <-
-    paste(
-      "drop table @synthea_schema.",
-      tablesToDrop,
-      ";",
-      collapse = "\n",
-      sep = ""
-    )
-  sql <- SqlRender::render(sql, synthea_schema = syntheaSchema)
-  sql <-
-    SqlRender::translate(sql, targetDialect = connectionDetails$dbms)
-  DatabaseConnector::executeSql(conn, sql)
+  
+  if (length(tablesToDrop) > 0) {
+      writeLines("Dropping Synthea tables...")
+	  sql <-
+		paste(
+		  "drop table @synthea_schema.",
+		  tablesToDrop,
+		  ";",
+		  collapse = "\n",
+		  sep = ""
+		)
+	  sql <- SqlRender::render(sql, synthea_schema = syntheaSchema)
+	  sql <- SqlRender::translate(sql, targetDialect = connectionDetails$dbms)
+	  DatabaseConnector::executeSql(conn, sql)
+  } else {
+	print(sprintf("No synthea tables to drop in schema %s",syntheaSchema))
+  }
+  
   on.exit(DatabaseConnector::disconnect(conn))
 }
